+++ aividades/atvteste/src/App.java
@@ -7,30 +7,30 @@ public class App {
 
         Scanner sc = new Scanner(System.in);
 
-        int codigo, qtd, valor, resul1, resul2, resultotal;
+        double codigo, qtd, valor, resul1, resul2, resultotal;
         
 
 
         System.out.println("Digite o codigo do produdo: ");
-                codigo = sc.nextInt();
+                codigo = sc.nextDouble();
         System.out.println("Digite a quantidade: ");
-                qtd = sc.nextInt();
+                qtd = sc.nextDouble();
         System.out.println("Digite valor: ");
-                valor = sc.nextInt();
+                valor = sc.nextDouble();
         
         resul1 = (valor * qtd);
 
         System.out.println("Digite o codigo do produdo: ");
-                codigo = sc.nextInt();
+                codigo = sc.nextDouble();
         System.out.println("Digite a quantidade: ");
-                qtd = sc.nextInt();
+                qtd = sc.nextDouble();
         System.out.println("Digite valor: ");
-                valor = sc.nextInt();
+                valor = sc.nextDouble();
 
         resul2 = (valor * qtd);
         resultotal = (resul1 + resul2);
 
-        System.out.printf("Valor total é : %d", resultotal);
+        System.out.printf("Valor total é : %.2f", resultotal);
